+++ test/font-families.test.js
@@ -3,11 +3,11 @@ const assert = require('assert')
 const vscode = require('vscode')
 const {
   _getFontFamiliesCache,
-  maybeShiftFontFamilyOnStartup,
-  getFontFamilies,
-  getCurrentFontFamily,
-  allFontFamilies,
+  shiftFontFamilyOnStartup,
+  getFontFamily,
   setFontFamily,
+  allFontFamilies,
+  getAvailableFontFamilies,
   DEFAULT_FONT_FAMILY,
 } = require('../src/font-families')
 const {
@@ -62,34 +62,34 @@ suite('font-families.test.js', () => {
       /\s/.test(ff.id),
     )
     await setFontFamily(fontFamilyWithSpace)
-    assert.strictEqual(getCurrentFontFamily(), fontFamilyWithSpace)
+    assert.strictEqual(getFontFamily(), fontFamilyWithSpace)
   })
 
   test('should not shift the font family when VS Code starts up if "shifty.startup.shiftFontFamilyOnStartup" is disabled', async () => {
-    await maybeShiftFontFamilyOnStartup()
-    assert.strictEqual(getCurrentFontFamily(), DEFAULT_FONT_FAMILY)
+    await shiftFontFamilyOnStartup()
+    assert.strictEqual(getFontFamily(), DEFAULT_FONT_FAMILY)
   })
 
   test('should shift the font family when VS Code starts up if "shifty.startup.shiftFontFamilyOnStartup" is enabled', async () => {
     await setConfig('shifty.startup.shiftFontFamilyOnStartup', true)
-    await maybeShiftFontFamilyOnStartup()
-    assert.notStrictEqual(getCurrentFontFamily(), DEFAULT_FONT_FAMILY)
+    await shiftFontFamilyOnStartup()
+    assert.notStrictEqual(getFontFamily(), DEFAULT_FONT_FAMILY)
   })
 
   test('should register font family commands when VS Code starts up', async () => {
     const commands = await vscode.commands.getCommands()
     assert.ok(commands.includes('shifty.shiftFontFamily'))
-    assert.ok(commands.includes('shifty.favoriteCurrentFontFamily'))
-    assert.ok(commands.includes('shifty.ignoreCurrentFontFamily'))
+    assert.ok(commands.includes('shifty.favoriteFontFamily'))
+    assert.ok(commands.includes('shifty.ignoreFontFamily'))
   })
 
   test('should shift the font family when running the "shifty.shiftFontFamily" command', async () => {
     await vscode.commands.executeCommand('shifty.shiftFontFamily')
-    assert.notStrictEqual(getCurrentFontFamily(), DEFAULT_FONT_FAMILY)
+    assert.notStrictEqual(getFontFamily(), DEFAULT_FONT_FAMILY)
   })
 
-  test('should favorite the current font family when running the "shifty.favoriteCurrentFontFamily" command', async () => {
-    await vscode.commands.executeCommand('shifty.favoriteCurrentFontFamily')
+  test('should favorite the current font family when running the "shifty.favoriteFontFamily" command', async () => {
+    await vscode.commands.executeCommand('shifty.favoriteFontFamily')
     assert.ok(
       getConfig('shifty.fontFamilies.favoriteFontFamilies').includes(
         DEFAULT_FONT_FAMILY,
@@ -101,17 +101,17 @@ suite('font-families.test.js', () => {
     )
   })
 
-  test('should ignore the current font family and shift the font family when running the "shifty.ignoreCurrentFontFamily" command', async () => {
-    await vscode.commands.executeCommand('shifty.ignoreCurrentFontFamily')
+  test('should ignore the current font family and shift the font family when running the "shifty.ignoreFontFamily" command', async () => {
+    await vscode.commands.executeCommand('shifty.ignoreFontFamily')
     assert.ok(
       getConfig('shifty.fontFamilies.ignoreFontFamilies').includes(
         DEFAULT_FONT_FAMILY,
       ),
     )
-    assert.notStrictEqual(getCurrentFontFamily(), DEFAULT_FONT_FAMILY)
+    assert.notStrictEqual(getFontFamily(), DEFAULT_FONT_FAMILY)
   })
 
-  test('should ignore the current font family and remove the font family from favorites when running the "shifty.ignoreCurrentFontFamily" command', async () => {
+  test('should ignore the current font family and remove the font family from favorites when running the "shifty.ignoreFontFamily" command', async () => {
     const favorites = [DEFAULT_FONT_FAMILY, 'SF Mono']
     await setConfig('shifty.fontFamilies.favoriteFontFamilies', favorites)
     assert.deepStrictEqual(
@@ -119,7 +119,7 @@ suite('font-families.test.js', () => {
       favorites,
     )
 
-    await vscode.commands.executeCommand('shifty.ignoreCurrentFontFamily')
+    await vscode.commands.executeCommand('shifty.ignoreFontFamily')
     assert.ok(
       getConfig('shifty.fontFamilies.ignoreFontFamilies').includes(
         DEFAULT_FONT_FAMILY,
@@ -144,7 +144,7 @@ suite('font-families.test.js', () => {
 
   test('should return all font families when no font families are ignored', () => {
     assert.strictEqual(
-      getFontFamilies().length,
+      getAvailableFontFamilies().length,
       allFontFamilies.filter(ff =>
         ff.supportedPlatforms.includes(DEFAULT_PLATFORM),
       ).length - 1,
@@ -152,18 +152,20 @@ suite('font-families.test.js', () => {
   })
 
   test('should return all font families except the current font family', () => {
-    assert.ok(getFontFamilies().every(ff => ff.id !== DEFAULT_FONT_FAMILY))
+    assert.ok(
+      getAvailableFontFamilies().every(ff => ff.id !== DEFAULT_FONT_FAMILY),
+    )
   })
 
   test('should return all font families except the ignored font families', async () => {
     const sfMono = 'SF Mono'
     await setConfig('shifty.fontFamilies.ignoreFontFamilies', [sfMono])
-    assert.ok(getFontFamilies().every(ff => ff.id !== sfMono))
+    assert.ok(getAvailableFontFamilies().every(ff => ff.id !== sfMono))
   })
 
   test('should return no codeface font families when ignored', async () => {
     await setConfig('shifty.fontFamilies.ignoreCodefaceFontFamilies', true)
-    assert.ok(getFontFamilies().every(ff => ff.type !== CODEFACE))
+    assert.ok(getAvailableFontFamilies().every(ff => ff.type !== CODEFACE))
   })
 
   test('should return font families that are supported on linux', async () => {
@@ -173,7 +175,9 @@ suite('font-families.test.js', () => {
     await setConfig('shifty.fontFamilies.fallbackFontFamily', '')
 
     assert.ok(
-      getFontFamilies().every(ff => ff.supportedPlatforms.includes(LINUX)),
+      getAvailableFontFamilies().every(ff =>
+        ff.supportedPlatforms.includes(LINUX),
+      ),
     )
   })
 
@@ -184,7 +188,9 @@ suite('font-families.test.js', () => {
     await setConfig('shifty.fontFamilies.fallbackFontFamily', '')
 
     assert.ok(
-      getFontFamilies().every(ff => ff.supportedPlatforms.includes(MAC_OS)),
+      getAvailableFontFamilies().every(ff =>
+        ff.supportedPlatforms.includes(MAC_OS),
+      ),
     )
   })
 
@@ -195,7 +201,9 @@ suite('font-families.test.js', () => {
     await setConfig('shifty.fontFamilies.fallbackFontFamily', '')
 
     assert.ok(
-      getFontFamilies().every(ff => ff.supportedPlatforms.includes(WINDOWS)),
+      getAvailableFontFamilies().every(ff =>
+        ff.supportedPlatforms.includes(WINDOWS),
+      ),
     )
   })
 
@@ -205,27 +213,27 @@ suite('font-families.test.js', () => {
     // change any shifty.fontFamilies config to reprime the cache
     await setConfig('shifty.fontFamilies.fallbackFontFamily', '')
 
-    assert.deepStrictEqual(getFontFamilies(), [DEFAULT_FONT_FAMILY])
+    assert.deepStrictEqual(getAvailableFontFamilies(), [DEFAULT_FONT_FAMILY])
   })
 
   test('should return user specified font families', async () => {
     const dankMono = 'Dank Mono'
     await setConfig('shifty.fontFamilies.includeFontFamilies', [dankMono])
-    assert.ok(getFontFamilies().find(ff => ff.id === dankMono))
+    assert.ok(getAvailableFontFamilies().find(ff => ff.id === dankMono))
   })
 
   test('should return favorite font familes when shiftMode is set to "favorites"', async () => {
     const favorites = ['Monaco', 'monofur', 'SF Mono']
     await setConfig('shifty.fontFamilies.favoriteFontFamilies', favorites)
     await setConfig('shifty.shiftMode', 'favorites')
-    assert.deepStrictEqual(getFontFamilies(), favorites)
+    assert.deepStrictEqual(getAvailableFontFamilies(), favorites)
   })
 
   test('should return font families without favorites when shiftMode is set to "discovery"', async () => {
     const sfMono = 'SF Mono'
     await setConfig('shifty.fontFamilies.favoriteFontFamilies', [sfMono])
     await setConfig('shifty.shiftMode', 'discovery')
-    assert.ok(!getFontFamilies().find(ff => ff.id === sfMono))
+    assert.ok(!getAvailableFontFamilies().find(ff => ff.id === sfMono))
   })
 
   test('should return favorite font families when shiftMode is set to "discovery" and all font families have been ignored or favorited', async () => {
@@ -235,11 +243,11 @@ suite('font-families.test.js', () => {
 
     // ignore the rest of the available font families
     await setConfig('shifty.fontFamilies.ignoreFontFamilies', [
-      ...getFontFamilies().map(ff => ff.id),
+      ...getAvailableFontFamilies().map(ff => ff.id),
       DEFAULT_FONT_FAMILY,
     ])
 
-    assert.deepStrictEqual(getFontFamilies(), favorites)
+    assert.deepStrictEqual(getAvailableFontFamilies(), favorites)
   })
 
   test('should return the default font family when shiftMode is set to "discovery" and all font families have been ignored', async () => {
@@ -247,10 +255,10 @@ suite('font-families.test.js', () => {
 
     // ignore all font families
     await setConfig('shifty.fontFamilies.ignoreFontFamilies', [
-      ...getFontFamilies().map(ff => ff.id),
+      ...getAvailableFontFamilies().map(ff => ff.id),
       DEFAULT_FONT_FAMILY,
     ])
 
-    assert.deepStrictEqual(getFontFamilies(), [DEFAULT_FONT_FAMILY])
+    assert.deepStrictEqual(getAvailableFontFamilies(), [DEFAULT_FONT_FAMILY])
   })
 })