+++ test/status-bar.test.js
@@ -15,11 +15,11 @@ suite('status-bar.test.js', () => {
 
   test('should register status bar commands when VS Code starts up', async () => {
     const commands = await vscode.commands.getCommands()
-    assert.ok(commands.includes('shifty.showCurrentStatus'))
+    assert.ok(commands.includes('shifty.showStatus'))
   })
 
-  test('should display the current color theme and font family when running the "shifty.showCurrentStatus" command', async () => {
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+  test('should display the current color theme and font family when running the "shifty.showStatus" command', async () => {
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.deepStrictEqual(
       vscode.window.showInformationMessage.secondCall.args,
       [
@@ -31,20 +31,20 @@ suite('status-bar.test.js', () => {
     )
   })
 
-  test('should display the stopped shift interval status when running the "shifty.showCurrentStatus" command', async () => {
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+  test('should display the stopped shift interval status when running the "shifty.showStatus" command', async () => {
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.deepStrictEqual(
       vscode.window.showInformationMessage.firstCall.args,
       ['Shift interval has not been started', 'Start shift interval'],
     )
   })
 
-  test('should display the color theme shift interval status when running the "shifty.showCurrentStatus" command', async () => {
+  test('should display the color theme shift interval status when running the "shifty.showStatus" command', async () => {
     await setConfig('shifty.shiftInterval.shiftColorThemeIntervalMs', 1337)
     await setConfig('shifty.shiftInterval.shiftFontFamilyIntervalMs', 0)
     await vscode.commands.executeCommand('shifty.startShiftInterval')
 
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.ok(
       vscode.window.showInformationMessage.firstCall.lastArg.match(
         /^\d{2}:\d{2} until color theme will shift$/,
@@ -54,12 +54,12 @@ suite('status-bar.test.js', () => {
     await vscode.commands.executeCommand('shifty.stopShiftInterval')
   })
 
-  test('should display the font family shift interval status when running the "shifty.showCurrentStatus" command', async () => {
+  test('should display the font family shift interval status when running the "shifty.showStatus" command', async () => {
     await setConfig('shifty.shiftInterval.shiftColorThemeIntervalMs', 0)
     await setConfig('shifty.shiftInterval.shiftFontFamilyIntervalMs', 1337)
     await vscode.commands.executeCommand('shifty.startShiftInterval')
 
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.ok(
       vscode.window.showInformationMessage.firstCall.lastArg.match(
         /^\d{2}:\d{2} until font family will shift$/,
@@ -69,12 +69,12 @@ suite('status-bar.test.js', () => {
     await vscode.commands.executeCommand('shifty.stopShiftInterval')
   })
 
-  test('should display the color theme and font family shift interval status when running the "shifty.showCurrentStatus" command', async () => {
+  test('should display the color theme and font family shift interval status when running the "shifty.showStatus" command', async () => {
     await setConfig('shifty.shiftInterval.shiftColorThemeIntervalMs', 1337)
     await setConfig('shifty.shiftInterval.shiftFontFamilyIntervalMs', 1337)
     await vscode.commands.executeCommand('shifty.startShiftInterval')
 
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.ok(
       vscode.window.showInformationMessage.firstCall.lastArg.match(
         /^\d{2}:\d{2} until color theme & font family will shift$/,
@@ -84,12 +84,12 @@ suite('status-bar.test.js', () => {
     await vscode.commands.executeCommand('shifty.stopShiftInterval')
   })
 
-  test('should display separate color theme and font family shift interval status when running the "shifty.showCurrentStatus" command', async () => {
+  test('should display separate color theme and font family shift interval status when running the "shifty.showStatus" command', async () => {
     await setConfig('shifty.shiftInterval.shiftColorThemeIntervalMs', 1337)
     await setConfig('shifty.shiftInterval.shiftFontFamilyIntervalMs', 7331)
     await vscode.commands.executeCommand('shifty.startShiftInterval')
 
-    await vscode.commands.executeCommand('shifty.showCurrentStatus')
+    await vscode.commands.executeCommand('shifty.showStatus')
     assert.ok(
       vscode.window.showInformationMessage.firstCall.lastArg.match(
         /^\d{2}:\d{2} until color theme will shift$/,